# openapi.fr.ex.custom.yaml
openapi: 3.0.3
info:
  title: SOAP REST Gateway API
  version: 1.0.1
  description: Spécification nettoyée (corruption retirée). À étendre progressivement.
servers:
  - url: /
paths:
...
  /api/create_contrat:
    post:
      summary: Création d'un contrat
      tags: [contrats]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/create_contratInput'
      responses:
        '200':
          description: Contrat créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContratSummary'
        '400':
          description: Erreur de validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
  /api/detail_contrat:
    post:
      summary: Détail d'un contrat
      tags: [contrats]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/detail_contratInput'
      responses:
        '200':
          description: Détail complet
          content:
            application/json:
              schema:
                type: object
                properties:
                  contrat:
                    $ref: '#/components/schemas/ContratSummary'
        '400':
          description: Erreur de validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
...
  /api/create_reglement:
    post:
      summary: Création de règlement
      tags: [reglements]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/create_reglementInput'
      responses:
        '200':
          description: Règlement enregistré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReglementSummary'
        '400':
          description: Erreur de validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erreur interne / Faute SOAP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SoapFaultError'
  /api/projects/create:
    post:
      summary: Création de projet
      tags: [projects]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/projects_project_createInput'
      responses:
        '200':
          description: Projet créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSummary'
        '400':
          description: Erreur de validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erreur interne / Faute SOAP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SoapFaultError'
  /api/projects/detail:
    post:
      summary: Détail projet
      tags: [projects]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/projects_project_detailInput'
      responses:
        '200':
          description: Détail du projet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSummary'
        '400':
          description: Erreur de validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erreur interne / Faute SOAP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SoapFaultError'
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
      required: [error]
    SoapFaultError:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            details:
              type: object
              properties:
                faultcode: { type: string }
                state: { type: string }
                errorCode: { type: string }
...
    ContratSummary:
      type: object
      properties:
        id: { type: number }
        produit: { type: string }
        effet: { type: string, format: date-time }
        statut: { type: string }
      required: [id, produit, effet]
...
    ReglementSummary:
      type: object
      properties:
        id: { type: number }
        typeoperation: { type: string }
        montant: { type: number }
        devise: { type: string }
        date: { type: string, format: date-time }
      required: [id, typeoperation, montant]
...
    ProjectSummary:
      type: object
      properties:
        idproj: { type: number }
        libelle: { type: string }
        username: { type: string }
      required: [idproj]
    ValidationErrorDetail:
      type: object
      properties:
        champ: { type: string, description: Nom du champ en erreur }
        message: { type: string, description: Raison de l’erreur }
      required: [champ, message]
    ValidationErrorResponse:
      type: object
      properties:
        erreur: { type: string, example: Le corps de la requête est invalide. }
        details:
          type: array
          items:
            $ref: '#/components/schemas/ValidationErrorDetail'
      required: [erreur, details]
    # Input Schemas
...
    detail_contratInput:
      type: object
      properties:
        contrat: { type: number }
        Allpieces: { type: boolean }
        DetailAdh: { type: boolean }
        Garanties: { type: boolean }
        Extensions: { type: boolean }
        infosCieProd: { type: boolean }
      required: [contrat, Allpieces, DetailAdh, Garanties, Extensions, infosCieProd]
...
    create_contratInput:
      type: object
      properties:
        dossier: { type: number }
        produit: { type: string }
        Effet: { type: string, format: date-time }
      required: [dossier, produit, Effet]
...
    create_reglementInput:
      type: object
      properties:
        typeoperation: { type: string }
        typeenc: { type: string }
        targetkind: { type: string }
        targetqintid: { type: string }
        montant: { type: number }
        devise: { type: string }
        date: { type: string, format: date-time }
        dateff: { type: string, format: date-time }
        reference: { type: string }
        tierspayeur: { type: string }
      required: [typeoperation, typeenc, targetkind, montant, devise]
...
    projects_project_createInput:
      type: object
      properties:
        dossier: { type: number }
        contrat: { type: number }
        produit: { type: string }
        username: { type: string }
        libelle: { type: string }
        resutXML: { type: boolean }
      required: [dossier, contrat, produit]
...
    projects_project_detailInput:
      type: object
      properties:
        dproj: { type: number }
      required: [dproj]
...